Moved to using a distinct dateformat in all SPG code.
