+++ src/main/java/uk/gov/ons/census/fwmt/outcomeservice/config/DateConfig.java
@@ -1,0 +1,12 @@
+package uk.gov.ons.census.fwmt.outcomeservice.config;
+
+import org.springframework.context.annotation.Bean;
+
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
+
+public class DateConfig {
+  @Bean DateFormat dateFormat() {
+    return new SimpleDateFormat("yyyy-MM-dd 'T'HH:mm:ss.SSSX");
+  }
+}
