+++ src/main/java/uk/gov/ons/census/fwmt/outcomeservice/message/RmFieldRepublishProducer.java
@@ -1,4 +1,4 @@
-package uk.gov.ons.census.fwmt.outcomeservice.service.impl;
+package uk.gov.ons.census.fwmt.outcomeservice.message;
 
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,7 +8,7 @@ import uk.gov.ons.census.fwmt.common.error.GatewayException;
 import uk.gov.ons.census.fwmt.common.rm.dto.FieldworkFollowup;
 
 @Service
-public class RmFieldRepublishService {
+public class RmFieldRepublishProducer {
 
   @Autowired
   @Qualifier("RmField")
